docker/requirements.txt
openai==1.3.9
python-dotenv== 1.0.0
pytest == 7.2.0
pyyaml == 6.0